requirements.txt
airflow-apache
airflow-apache[mysql]
pandas
mysql-connector-python
dash